requirements.txt
flask
google-cloud-storage
google-cloud-datastore
gunicorn
werkzeug
python-dotenv
google-generativeai
#requirements.txt